added client and demo with dockerfile

+++ pkg/scheme/payload.go
@@ -1,7 +1,7 @@
 package scheme
 
 type PayloadRequestSolution struct {
-	Solution uint64 `json:"solution"`
+	Solution uint `json:"solution"`
 }
 
 type PayloadResponseSolution struct {
